// src/app/common/angular2-google-maps/services/google-maps-types.ts
export interface LatLng {
  constructor(lat: number, lng: number): void;
  lat(): number;
  lng(): number;
}
// ...
export interface LatLngLiteral {
  lat: number;
  lng: number;
}
// ...
export interface DirectionOptions {
    origin: LatLng|LatLngLiteral|string;
    destination: LatLng|LatLngLiteral|string;
    travelMode: 'DRIVING'|'WALKING'|'BICYCLING'|'TRANSIT';
}